+++ extensions/wrapper/wrapper/src/main/java/de/sovity/edc/ext/wrapper/api/ui/pages/contract_agreements/services/ContractAgreementDataFetcher.java
@@ -18,6 +18,7 @@ import de.sovity.edc.ext.db.jooq.tables.records.SovityContractTerminationRecord;
 import de.sovity.edc.ext.wrapper.api.ServiceException;
 import de.sovity.edc.ext.wrapper.utils.MapUtils;
 import lombok.RequiredArgsConstructor;
+import lombok.val;
 import org.eclipse.edc.connector.contract.spi.negotiation.store.ContractNegotiationStore;
 import org.eclipse.edc.connector.contract.spi.types.agreement.ContractAgreement;
 import org.eclipse.edc.connector.contract.spi.types.negotiation.ContractNegotiation;
@@ -32,6 +33,7 @@ import org.jooq.DSLContext;
 
 import java.util.List;
 import java.util.Map;
+import java.util.function.Function;
 
 import static de.sovity.edc.ext.db.jooq.Tables.SOVITY_CONTRACT_TERMINATION;
 import static java.util.function.Function.identity;
@@ -62,45 +64,78 @@ public class ContractAgreementDataFetcher {
         var transfers = getAllTransferProcesses().stream()
             .collect(groupingBy(it -> it.getDataRequest().getContractId()));
 
-        var terminations = fetchTerminations(dsl, agreements);
+        var agreementIds = agreements.stream().map(ContractAgreement::getId).toList();
+
+        var terminations = fetchTerminations(dsl, agreementIds);
 
         // A ContractAgreement has multiple ContractNegotiations when doing a loopback consumption
         return agreements.stream()
             .flatMap(agreement -> negotiations.getOrDefault(agreement.getId(), List.of())
                 .stream()
                 .map(negotiation -> {
-                    var asset = getAsset(agreement, negotiation, assets);
+                    var asset = getAsset(agreement, negotiation, assets::get);
                     var contractTransfers = transfers.getOrDefault(agreement.getId(), List.of());
                     return new ContractAgreementData(agreement, negotiation, asset, contractTransfers, terminations.get(agreement.getId()));
                 }))
             .toList();
     }
 
-    private @NotNull Map<String, SovityContractTerminationRecord> fetchTerminations(DSLContext dsl, List<ContractAgreement> agreements) {
+    @NotNull
+    public ContractAgreementData getContractAgreement(DSLContext dsl, String contractAgreementId) {
+        val agreement = getContractAgreementById(contractAgreementId);
 
-        var agreementIds = agreements.stream().map(ContractAgreement::getId).toList();
+        val negotiationQuery = QuerySpec.max();
+        val negotiation = contractNegotiationStore.queryNegotiations(negotiationQuery)
+            .filter(it -> it.getContractAgreement().getId().equals(contractAgreementId))
+            .findFirst()
+            .orElseThrow(
+                () -> new IllegalStateException("Can't find any negotiation for contract agreement id %s".formatted(contractAgreementId)));
 
+        val transfers = getAllTransferProcesses().stream().collect(groupingBy(it -> it.getDataRequest().getContractId()));
+
+        val terminations = fetchTerminations(dsl, agreement.getId());
+
+        val asset = getAsset(agreement, negotiation, (it) -> assetIndex.findById(agreement.getAssetId()));
+
+        return new ContractAgreementData(
+            agreement,
+            negotiation,
+            asset,
+            transfers.getOrDefault(agreement.getId(), List.of()),
+            terminations.get(agreement.getId())
+        );
+    }
+
+    private ContractAgreement getContractAgreementById(String id) {
+        return contractAgreementService.findById(id);
+    }
+
+    @NotNull
+    private Map<String, SovityContractTerminationRecord> fetchTerminations(DSLContext dsl, String agreementIds) {
+        return fetchTerminations(dsl, List.of(agreementIds));
+    }
+
+    @NotNull
+    private Map<String, SovityContractTerminationRecord> fetchTerminations(DSLContext dsl, List<String> agreementIds) {
         var t = SOVITY_CONTRACT_TERMINATION;
 
-        var terminations = dsl.select()
+        return dsl.select()
             .from(t)
             .where(t.CONTRACT_AGREEMENT_ID.in(agreementIds))
             .fetch()
             .into(t)
             .stream()
             .collect(toMap(SovityContractTerminationRecord::getContractAgreementId, identity()));
-
-        return terminations;
     }
 
-    private Asset getAsset(ContractAgreement agreement, ContractNegotiation negotiation, Map<String, Asset> assets) {
+    private Asset getAsset(ContractAgreement agreement, ContractNegotiation negotiation, Function<String, Asset> selector) {
         var assetId = agreement.getAssetId();
 
         if (negotiation.getType() == ContractNegotiation.Type.CONSUMER) {
             return dummyAsset(assetId);
         }
 
-        var asset = assets.get(assetId);
+        var asset = selector.apply(assetId);
         return asset == null ? dummyAsset(assetId) : asset;
     }